refactor(core): standardize Symbol DI token re-exports across contexts
Make the Symbol-token publication pattern uniform across every bounded
context in `libs/core/src`. Before: 9 sub-barrels cherry-picked a hand-
maintained subset of token names from `<ctx>.tokens.ts` (sync, products,
inventory, orders, integrations, mappings, listings, events) or from
the Nest module file (users), and one context declared its token inline
in a port file (webhooks). After: every context owns a token-only
`<ctx>/<ctx>.tokens.ts` and the sub-barrel does `export * from './<ctx>.tokens'`,
so adding a new token in `<ctx>.tokens.ts` is automatically published —
no second edit required, no risk of barrels drifting behind.

Changes:

- Add `libs/core/src/users/users.tokens.ts` factoring the 3 inline
  declarations out of `users.module.ts`.
- Add `libs/core/src/webhooks/webhooks.tokens.ts` factoring
  `WEBHOOK_DELIVERY_REPOSITORY_TOKEN` out of the port file.
- Normalize 9 sub-barrels (events, integrations, inventory, listings,
  mappings, orders, products, sync, users, webhooks) to
  `export * from './<ctx>.tokens'`.
- Rewrite 12 cross-context deep imports of `@openlinker/core/<ctx>/<ctx>.tokens`
  to the top-level barrel `@openlinker/core/<ctx>` (3 in libs/core,
  1 in apps/api, 7 in apps/worker integration specs, 1 in libs/core
  integrations.service.spec).
- Block `@openlinker/core/*/*.tokens` deep imports via three
  `no-restricted-imports` patterns in `.eslintrc.js` (port files,
  integration packages, backend host apps).
- Add `### Symbol DI Token Re-export Convention` section to
  `docs/engineering-standards.md` codifying the 6 rules.
- Add the implementation plan at
  `docs/plans/implementation-plan-symbol-token-reexports.md`.

Quality gate clean: lint 0 errors, type-check passes, 1826 unit tests
green across all backend + plugin packages, 1064 vitest tests green
in apps/web.

Closes #595
