+++ .eslintrc.js
@@ -414,10 +414,11 @@ module.exports = {
                   '@openlinker/core/*/application/**',
                   '@openlinker/core/*/infrastructure/**',
                   '@openlinker/core/*/orm-entities',
+                  '@openlinker/core/*/*.tokens',
                   '@openlinker/integrations-*/**',
                 ],
                 message:
-                  "Port and capability files must import cross-context types via the top-level package barrel — e.g. `import { Connection } from '@openlinker/core/identifier-mapping'` — never via deep sub-paths. ORM-entity sub-barrels (`@openlinker/core/*/orm-entities`) are host-only — they must not appear on a port's import list so the contract surface stays framework-neutral (#594). Integration-package imports invert the dependency direction.",
+                  "Port and capability files must import cross-context types via the top-level package barrel — e.g. `import { Connection } from '@openlinker/core/identifier-mapping'` — never via deep sub-paths. ORM-entity sub-barrels (`@openlinker/core/*/orm-entities`) are host-only — they must not appear on a port's import list so the contract surface stays framework-neutral (#594). Symbol DI tokens (`@openlinker/core/*/*.tokens`) are re-exported from the top-level context barrel; deep paths are forbidden (#595). Integration-package imports invert the dependency direction.",
               },
             ],
           },
@@ -445,9 +446,10 @@ module.exports = {
                   '@openlinker/core/*/application/**',
                   '@openlinker/core/*/infrastructure/**',
                   '@openlinker/core/*/orm-entities',
+                  '@openlinker/core/*/*.tokens',
                 ],
                 message:
-                  'Integration packages must import from `@openlinker/core/<context>` top-level barrels — never deep sub-paths or ORM-entity sub-barrels. Deep imports leak unstable internals; ORM-entity imports couple the plugin to TypeORM. See #591 and #594.',
+                  'Integration packages must import from `@openlinker/core/<context>` top-level barrels — never deep sub-paths, ORM-entity sub-barrels, or `*.tokens` paths. Deep imports leak unstable internals; ORM-entity imports couple the plugin to TypeORM; deep token paths fragment the DI surface. See #591, #594, and #595.',
               },
             ],
           },
@@ -473,9 +475,10 @@ module.exports = {
                   '@openlinker/core/*/domain/**',
                   '@openlinker/core/*/application/**',
                   '@openlinker/core/*/infrastructure/**',
+                  '@openlinker/core/*/*.tokens',
                 ],
                 message:
-                  'Backend apps must import from `@openlinker/core/<context>` top-level barrels — never deep sub-paths. The package.json wildcards were dropped in #591; deep aliases now fail at Node runtime.',
+                  'Backend apps must import from `@openlinker/core/<context>` top-level barrels — never deep sub-paths. The package.json wildcards were dropped in #591; deep aliases now fail at Node runtime. Symbol DI tokens are re-exported from the top-level barrel (#595).',
               },
             ],
           },
