+++ libs/core/src/integrations/application/services/integrations.service.spec.ts
@@ -13,8 +13,8 @@ import { AdapterRegistryPort } from '../../domain/ports/adapter-registry.port';
 import { IdentifierMappingPort } from '@openlinker/core/identifier-mapping';
 import { CredentialsResolverPort } from '../../domain/ports/credentials-resolver.port';
 import { AdapterFactoryResolverService } from '../../infrastructure/adapters/adapter-factory-resolver.service';
-import { CONNECTION_PORT_TOKEN, IDENTIFIER_MAPPING_PORT_TOKEN } from '@openlinker/core/identifier-mapping/identifier-mapping.tokens';
-import { ADAPTER_REGISTRY_TOKEN, ADAPTER_FACTORY_RESOLVER_TOKEN, CREDENTIALS_RESOLVER_TOKEN } from '@openlinker/core/integrations/integrations.tokens';
+import { CONNECTION_PORT_TOKEN, IDENTIFIER_MAPPING_PORT_TOKEN } from '@openlinker/core/identifier-mapping';
+import { ADAPTER_REGISTRY_TOKEN, ADAPTER_FACTORY_RESOLVER_TOKEN, CREDENTIALS_RESOLVER_TOKEN } from '@openlinker/core/integrations';
 import { Connection } from '@openlinker/core/identifier-mapping';
 import { ConnectionNotFoundException } from '@openlinker/core/identifier-mapping';
 import { ConnectionDisabledException } from '@openlinker/core/identifier-mapping';
