+++ libs/core/src/webhooks/webhooks.tokens.ts
@@ -1,0 +1,9 @@
+/**
+ * Webhooks DI Tokens
+ *
+ * Symbol tokens for dependency injection in the webhooks bounded context.
+ *
+ * @module libs/core/src/webhooks
+ */
+
+export const WEBHOOK_DELIVERY_REPOSITORY_TOKEN = Symbol('WebhookDeliveryRepositoryPort');
